// src/public_pages/carrier/job_application/model/data/IPersonSectionState.ts
export interface IPersonSectionState {
  ApplicantId: string;
  ApplicantPhoto: string;
  ApplicantFullName: string;
  FatherName: string;
  MotherName: string;
  Gender: string;
  NationalityId: number;
  NationalityName: string;
  DateOfBirth: string;
  ReligionId: number;
  ReligionName: string;
  NationalIdNumber: string;
  ETinNumber: string;
  PassportNumber: string;
  BloodGroup: string;
  ApplicantEmergencyContactNumber: string;
  ApplicantEmergencyContactName: string;
  ApplicantEmergencyContactRelationId: number;
  ApplicantEmergencyContactRelationName: string;
  MaritalStatus: string;
  SpouseName: string;
  PresentAddress1: string;
  PresentAddress2: string;
  PresentDivisionId: number;
  PresentAddress3: number;
  SubDistrictName: string;
  PresentDivisionName: string;
  PresentAddress4: number;
  PresentDistrictName: string;
  PresentAddress5: number;
  PresentVillageName: string;
  PermanentAddress1: string;
  PermanentAddress2: string;
  PermanentDivisionId: number;
  PermanentDivisionName: string;
  PermanentAddress3: number;
  PermanentPoliceStation: string;
  PermanentAddress4: number;
  PermanentDistrictName: number;
  PermanentAddress5: number;
  PermanentVillage: string;
  ApplicantPersonalEmail: string;
  ApplicantOfficialEmail: string;
  ApplicantCellNumber: string;
  ApplicantWorkNumber: string;
  ApplicantHomeNumber: string;
  ApplicantGender: string;
  ApplicantMaritalStatus: string;
  Errors: {
    ApplicantId: string;
    ApplicantPhoto: string;
    ApplicantFullName: string;
    FatherName: string;
    MotherName: string;
    Gender: string;
    NationalityId: string;
    DateOfBirth: string;
    ReligionId: string;
    NationalIdNumber: string;
    ETinNumber: string;
    PassportNumber: string;
    BloodGroup: string;
    ApplicantEmergencyContactNumber: string;
    ApplicantEmergencyContactName: string;
    ApplicantEmergencyContactRelationId: string;
    MaritalStatus: string;
    SpouseName: string;
    PresentAddress1: string;
    PresentAddress2: string;
    PresentDivision: string;
    PresentAddress3: string;
    PresentAddress4: string;
    PresentAddress5: string;
    PermanentAddress1: string;
    PermanentAddress2: string;
    PermanentDivision: string;
    PermanentAddress3: string;
    PermanentAddress4: string;
    PermanentAddress5: string;
    ApplicantPersonalEmail: string;
    ApplicantOfficialEmail: string;
    ApplicantCellNumber: string;
    ApplicantWorkNumber: string;
    ApplicantHomeNumber: string;
    ApplicantGender: string;
    ReligionName: string;
    NationalityName: string;
    ApplicantMaritalStatus: string;
    ApplicantEmergencyContactRelationName: string;
  };
}